WIP - Moving core data logic out of app delegate
